+++ Mods/S7_UI_Components_Library_b66d56c6-12f9-4abc-844f-0c30b89d32e4/Story/RawFiles/Lua/Components/Journal.lua
@@ -216,6 +216,22 @@ local function RegisterJournalListeners()
         })
     end)
 
+    --  UPDATE TEXT
+    --  -----------
+
+    Ext.RegisterUICall(Journal.UI, "textUpdate", function (ui, call, id, updatedText)
+        local catMapID, chapMapID, paraMapID, journalNodeType = parseID(id)
+        local Pos, catPos, chapPos, paraPos = getPos(id)
+
+        if journalNodeType == 1 then
+            Journal.JournalData[catPos]["strContent"] = updatedText
+        elseif journalNodeType == 2 then
+            Journal.JournalData[catPos]["chapters"][chapPos]["strContent"] = updatedText
+        elseif journalNodeType == 3 then
+            Journal.JournalData[catPos]["chapters"][chapPos]["paragraphs"][paraPos]["strContent"] = updatedText
+        end
+    end)
+
     --  REMOVE NODES
     --  ============